+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [1.2]
+### - 2024-06-19
+- out 패키지에 JsonTransformer 클래스 추가 jsonString 으로 변환하되 numberic 타입은 numberic 타입으로 자동 형변화 기능
 
 ### - 2024-05-09
 - HttpRequest v1.0.2 -> v1.0.3, set 함수 추가 (생성자에서 배열로 등록해야만 했던 요청 주소및파라메터들을 쉽게 등록할 수 있도록 업데이트)